requirements.txt
requests==2.32.3
retry==0.9.2
python-dotenv==1.0.1